ajout de l'action delete

+++ Bureau/Symfony/projet_final/src/AppBundle/Controller/MuseeController.php
@@ -7,6 +7,7 @@ use AppBundle\Entity\Commentaire;
 use AppBundle\Entity\User;
 use AppBundle\Repository;
 use Doctrine\DBAL\Types\StringType;
+use phpDocumentor\Reflection\Types\Integer;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\BrowserKit\Response;
@@ -114,7 +115,7 @@ class MuseeController extends Controller
 
     public function FormOfMusee(Request $request)
     {
-
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'Unable to access this page!');
         ///Formulaire des Musees
         $musee = new Musee();
         $form = $this->createForm('AppBundle\Form\MuseeType', $musee);
@@ -138,18 +139,43 @@ class MuseeController extends Controller
 
 
     /**
-     * @Route("/supprim/{musee}" , name="supprission")
+     * @Route("/supprim/{id}" , name="supprission")
      */
-    public function supprimMusee($musee){
-
+    public function supprimAction( $id){
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'Unable to access this page!');
 
         $em = $this->getDoctrine()->getManager();
+        $musee =  $em->getRepository('AppBundle:Musee')->find($id);
         $em->remove($musee);
         $em->flush();
         return $this->redirectToRoute('musee', array('page' => 108));
 
     }
 
+
+    /**
+     * @Route("/edit/{id}" , name="modifier")
+     */
+
+    public function EditMuseeAction(Musee $musee, Request $request)
+    {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'Unable to access this page!');
+        $form = $this->createForm('AppBundle\Form\MuseeType', $musee);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $manager = $this->getDoctrine()->getManager();
+
+            $manager->flush();
+
+            //Retour vers la page d'acceuille
+            return $this->redirectToRoute('start');
+        }
+
+        return $this->render(array ('musee/ModifieMusee.html.twig',
+            'form' => $form->createView(),
+        ));
+    }
 
 }
